requirements.txt
# Core dependencies
python-dotenv
streamlit
langchain
langchain-community
langchain-core
langchain-experimental
langchain-text-splitters
langchain-chroma
langchain-google-genai
langchain_groq
langgraph
fastapi
uvicorn
unstructured
pydantic

# Document processing
pyPDF2
python-pptx
python-docx
pandas
markdown

# Vector database
chromadb

# Embedding models
google-generativeai

# Utilities
tqdm
requests
uuid